update common location + test
